Restuctured code to new files and refactored the panel code
