+++ src/config.js
@@ -1,0 +1,6 @@
+const ALARM_NAMESPACE = 'tomatoClockAlarm';
+const NOTIFICATION_ID = 'tomatoClockNotification';
+const STORAGE_KEY = 'timeline';
+const MINUTES_IN_TOMATO = 25;
+const MINUTES_IN_SHORT_BREAK = 5;
+const MINUTES_IN_LONG_BREAK = 15;
