+++ src/panel/panel.html
@@ -28,6 +28,9 @@
 	</div>
 
 	<script src="/lib/browser-polyfill.js"></script>
+	<script src="../config.js"></script>
+	<script src="../utils.js"></script>
+	<script src="../models/timer.js"></script>
 	<script src="panel.js"></script>
 </body>
 </html>
