+++ src/panel/panel.js
@@ -1,58 +1,58 @@
-const ALARM_NAMESPACE = 'tomatoClockAlarm';
-
-function getSecondsInMilliseconds(seconds) {
-	return seconds * 1000;
-}
-function getMinutesInMilliseconds(minutes) {
-	return minutes * 60000;
-}
-
-function millisecondsToTimeText(milliseconds) {
-	const minutes = parseInt((milliseconds / (1000 * 60)) % 60);
-	const seconds = parseInt((milliseconds / 1000) % 60);
-	const minutesString = (minutes < 10) ? `0${minutes}` : minutes.toString();
-	const secondsString = (seconds < 10) ? `0${seconds}` : seconds.toString();
-
-	return `${minutesString}:${secondsString}`;
-}
-
-
-
 class Panel {
+	
 	constructor() {
 		this.currentTimeText = document.getElementById('current-time-text');
-		this.timer = {};
-
+		this.timer = new Timer();
+		
+		var _this = this;
 		browser.runtime.sendMessage({
-			action: 'getTimerScheduledTime'
-		}).then(scheduledTime => {
-			if (scheduledTime) {
-				this.setTimer(scheduledTime - Date.now());
+			action: 'getBackgroundTimer'
+		}).then(timer => {
+			if (timer) {
+				timer = JSON.parse(timer);
+				this.timer = new Timer(timer);
+				this.timer.registerStartedEventHandler(_this);
+				this.timer.registerUpdatedEventHandler(_this);
+				this.timer.registerFinishedEventHandler(_this);
+				this.timer.registerCanceledEventHandler(_this);
+				this.onTimerStarted(this.timer);
 			}
 		});
 
 		this.setEventListeners();
 	}
+	
+	onTimerStarted(timer) {
+		this.setCurrentTimeText(timer.timeLeft);
+	}
+	
+	onTimerUpdated(timer) {
+		this.setCurrentTimeText(timer.timeLeft);
+	}
+	
+	onTimerFinished(timer) {
+		this.setCurrentTimeText(0);
+	}
+	
+	onTimerCanceled(timer) {
+		this.setCurrentTimeText(0);
+	}
 
 	setEventListeners() {
 		document.getElementById('tomato-button').addEventListener('click', () => {
-			this.setTimer(getMinutesInMilliseconds(25));
-			this.setBackgroundTimer(getMinutesInMilliseconds(25));
+			this.setTimers(getMinutesInMilliseconds(MINUTES_IN_TOMATO));
 		});
 
 		document.getElementById('short-break-button').addEventListener('click', () => {
-			this.setTimer(getMinutesInMilliseconds(5));
-			this.setBackgroundTimer(getMinutesInMilliseconds(5));
+			this.setTimers(getMinutesInMilliseconds(MINUTES_IN_SHORT_BREAK));
 		});
 
 		document.getElementById('long-break-button').addEventListener('click', () => {
-			this.setTimer(getMinutesInMilliseconds(15));
-			this.setBackgroundTimer(getMinutesInMilliseconds(15));
+			this.setTimers(getMinutesInMilliseconds(MINUTES_IN_LONG_BREAK));
 		});
 
 		document.getElementById('reset-button').addEventListener('click', () => {
-			this.resetTimer();
-			this.resetBackgroundTimer();
+			this.resetTimers();
 		});
 
 		document.getElementById('stats-link').addEventListener('click', () => {
@@ -60,53 +60,21 @@ class Panel {
 		});
 	}
 
-	resetTimer() {
-		clearInterval(this.timer.interval);
-
-		this.timer = {
-			interval: null,
-			timeLeft: 0
-		}
-
-		this.setCurrentTimeText(0);
-	}
-
-	getTimer() {
-		return this.timer;
-	}
-
-	setTimer(milliseconds) {
-		this.resetTimer();
-		this.setCurrentTimeText(milliseconds);
-
-		this.timer = {
-			interval: setInterval(() => {
-				const timer = this.getTimer();
-
-				timer.timeLeft -= getSecondsInMilliseconds(1);
-				this.setCurrentTimeText(timer.timeLeft);
-
-				if (timer.timeLeft <= 0) {
-					this.resetTimer();
-				}
-			}, getSecondsInMilliseconds(1)),
-			timeLeft: milliseconds
-		};
-	}
-
 	setCurrentTimeText(milliseconds) {
 		this.currentTimeText.textContent = millisecondsToTimeText(milliseconds);
 	}
 
-	resetBackgroundTimer() {
+	resetTimers() {
+		this.timer.reset();
 		browser.runtime.sendMessage({
 			action: 'resetTimer'
-		});
+		}); 
 	}
 
-	setBackgroundTimer(milliseconds) {
+	setTimers(milliseconds) {
 		const minutes = milliseconds / 60000;
 
+		this.timer.set(milliseconds);
 		browser.runtime.sendMessage({
 			action: 'setTimer',
 			data: {
