+++ src/utils.js
@@ -1,0 +1,22 @@
+function getSecondsInMilliseconds(seconds) {
+	return seconds * 1000;
+}
+function getMinutesInMilliseconds(minutes) {
+	return minutes * 60000;
+}
+
+function millisecondsToMinutesAndSeconds(milliseconds) {
+	return {
+		minutes: parseInt((milliseconds / (1000 * 60)) % 60),
+		seconds: parseInt((milliseconds / 1000) % 60)
+	};
+}
+
+function millisecondsToTimeText(milliseconds) {
+	const minutes = parseInt((milliseconds / (1000 * 60)) % 60);
+	const seconds = parseInt((milliseconds / 1000) % 60);
+	const minutesString = (minutes < 10) ? `0${minutes}` : minutes.toString();
+	const secondsString = (seconds < 10) ? `0${seconds}` : seconds.toString();
+
+	return `${minutesString}:${secondsString}`;
+}
